requirements.txt
bcrypt
Django
django-anymail
django-cors-headers
django-environ
django-hashid-field
django-model-utils
django-rest-registration
djangorestframework
djangorestframework-simplejwt
drf-spectacular
PyJWT
pyotp
twilio
pydantic
django-cors-headers
django-debug-toolbar